refactor(store): move get issues from repos to util
