+++ lib/store/IssueStore.js
@@ -6,15 +6,6 @@ import parseURL from "github-url-parse"
 import storage from "local-storage"
 import GitHubAPI from "../util/GitHubAPI"
 import Promise from "bluebird"
-function flatten([first, ...rest]) {
-    if (first === undefined) {
-        return [];
-    } else if (!Array.isArray(first)) {
-        return [first, ...flatten(rest)];
-    } else {
-        return [...flatten(first), ...flatten(rest)];
-    }
-}
 function sortDate(aItem, bItem) {
     var a = new Date(aItem["updated_at"]),
         b = new Date(bItem["updated_at"]);
@@ -38,32 +29,32 @@ export default class IssueStore extends Store {
         this.register(keys.removeFavRepository, this.onRemoveFavIssue.bind(this));
         this.register(keys.validGitHubToken, this.onValidGitHubToken.bind(this));
 
-        this.fetchIssuesFromRepositories(this.state.favRepositories);
+        this.loadIssuesFromRepositories(this.state.favRepositories);
     }
-    
+
     getMyIssues() {
         return this.state.myIssues;
     }
-    
+
     getRepoIssues() {
         return this.state.repoIssues;
     }
-    
+
     getViewingIssue() {
         return this.state.viewingIssue;
     }
-    
+
     getViewingIssueComments() {
         return this.state.viewingIssueComments;
     }
-    
+
     getViewingURL() {
         if (this.state.viewingIssue == null) {
             return "https://github.com/";
         }
         return this.state.viewingIssue.html_url;
     }
-    
+
     isFavIssue(issue) {
         if (issue == null) {
             return false;
@@ -75,18 +66,14 @@ export default class IssueStore extends Store {
             return favRepository === repositoryURL;
         });
     }
-    
-    fetchIssuesFromRepositories(repositories) {
+
+    loadIssuesFromRepositories(repositories) {
         var api = new GitHubAPI();
-        var promises = repositories.map((repository)=> {
-            var ghObject = parseURL(repository);
-            return api.getRepoIssuesAsync(ghObject.user, ghObject.repo);
-        });
-        Promise.all(promises).then((issues) => {
+        return api.getIssuesFromRepositories(repositories).then((issues)=> {
             this.setState({
-                repoIssues: flatten(issues).sort(sortDate)
+                repoIssues: issues.sort(sortDate)
             });
-        });
+        })
     }
 
     isValidGitHubToken() {
@@ -107,12 +94,12 @@ export default class IssueStore extends Store {
         }
         repositories.push(repositoryURL);
         storage.set("favRepositories", repositories);
-        this.fetchIssuesFromRepositories(repositories);
+        this.loadIssuesFromRepositories(repositories);
         this.setState({
             favRepositories: repositories
         });
     }
-    
+
     onRemoveFavIssue(issue) {
         var ghObject = parseURL(issue.html_url);
         var repositoryURL = "https://github.com/" + ghObject.user + "/" + ghObject.repo;
@@ -128,24 +115,24 @@ export default class IssueStore extends Store {
             return favRepository !== repositoryURL;
         });
         storage.set("favRepositories", filteredRepositories);
-        this.fetchIssuesFromRepositories(filteredRepositories);
+        this.loadIssuesFromRepositories(filteredRepositories);
         this.setState({
             favRepositories: filteredRepositories
         });
     }
-    
+
     onLoadMyIssues(issues) {
         this.setState({
             myIssues: issues
         });
     }
-    
+
     onLoadRepoIssues(issues) {
         this.setState({
             repoIssues: issues
         });
     }
-    
+
     onViewIssue(issue, comments) {
         this.setState({
             viewingIssue: issue,
