+++ lib/util/GitHubAPI.js
@@ -2,6 +2,17 @@
 "use strict";
 import GitHub from "github"
 import Promise from "bluebird"
+import parseURL from "github-url-parse"
+function flatten([first, ...rest]) {
+    if (first === undefined) {
+        return [];
+    } else if (!Array.isArray(first)) {
+        return [first, ...flatten(rest)];
+    } else {
+        return [...flatten(first), ...flatten(rest)];
+    }
+}
+
 class GitHubAPI {
     constructor(token) {
         this.client = new GitHub({
@@ -38,6 +49,14 @@ class GitHubAPI {
                 }
             });
         });
+    }
+
+    getIssuesFromRepositories(repositories) {
+        var promises = repositories.map((repository)=> {
+            var ghObject = parseURL(repository);
+            return this.getRepoIssuesAsync(ghObject.user, ghObject.repo);
+        });
+        return Promise.all(promises);
     }
 
     getRepoIssuesAsync(user, repo) {
